* grafana/master: (30 commits)
  Plugins: update beta notice style (#16928)
  Chore: update version number for 6.3 (#16927)
  Plugins: Support templated urls in routes (#16599)
  changelog: add 5.4.4 release
  docs: add download link to what's new in v6.2
  update changelog
  Update changelog for 6.2.0-beta1
  AzureMonitor: adds support for multiple subscriptions per datasource (#16922)
  docs: what's new in v6.2 (#16909)
  Chore: ban importing from @grafana/ui in grafana ui files (#16920)
  BarGauge: New multi series enabled gauge like panel with horizontal and vertical layouts and 3 display modes (#16918)
  alerting: no notification when going from nodata -> pending (#16905)
  rpm: start grafana after mysqld process (#16917)
  Build: fix failing grafana/ui build (#16913)
  BarGauge: Updated test dashboards and threshold fix (#16911)
  PanelModel: Clear queryRunner on destroy (#16906)
  Panels: No title will no longer make panel header take up space (#16884)
  Elasticsearch: Fix pre-v7.0 and alerting error (#16904)
  Gauge: Better handling of gauge repeat title (#16886)
  Refactor: move datemath to grafana/ui (#16890)
  ...

+++ pkg/tsdb/stackdriver/stackdriver_test.go
@@ -236,13 +236,13 @@ func TestStackdriver(t *testing.T) {
 
 				Convey("timestamps should be in ascending order", func() {
 					So(res.Series[0].Points[0][0].Float64, ShouldEqual, 0.05)
-					So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1536670020000)
+					So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1536670020000))
 
 					So(res.Series[0].Points[1][0].Float64, ShouldEqual, 1.05)
-					So(res.Series[0].Points[1][1].Float64, ShouldEqual, 1536670080000)
+					So(res.Series[0].Points[1][1].Float64, ShouldEqual, int64(1536670080000))
 
 					So(res.Series[0].Points[2][0].Float64, ShouldEqual, 1.0666666666667)
-					So(res.Series[0].Points[2][1].Float64, ShouldEqual, 1536670260000)
+					So(res.Series[0].Points[2][1].Float64, ShouldEqual, int64(1536670260000))
 				})
 			})
 
@@ -365,9 +365,9 @@ func TestStackdriver(t *testing.T) {
 				}
 
 				Convey("timestamps should be in ascending order", func() {
-					So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1536668940000)
-					So(res.Series[0].Points[1][1].Float64, ShouldEqual, 1536669000000)
-					So(res.Series[0].Points[2][1].Float64, ShouldEqual, 1536669060000)
+					So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1536668940000))
+					So(res.Series[0].Points[1][1].Float64, ShouldEqual, int64(1536669000000))
+					So(res.Series[0].Points[2][1].Float64, ShouldEqual, int64(1536669060000))
 				})
 
 				Convey("bucket bounds should be correct", func() {
@@ -410,8 +410,8 @@ func TestStackdriver(t *testing.T) {
 				}
 
 				Convey("timestamps should be in ascending order", func() {
-					So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1550859086000)
-					So(res.Series[0].Points[1][1].Float64, ShouldEqual, 1550859146000)
+					So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1550859086000))
+					So(res.Series[0].Points[1][1].Float64, ShouldEqual, int64(1550859146000))
 				})
 
 				Convey("bucket bounds should be correct", func() {
